BookingService and interface added
