+++ app/models/IBooking.ts
@@ -1,0 +1,8 @@
+export interface IBooking {
+  id: string,
+  restaurantId: string,
+  date: string,
+  time: string,
+  numberOfGuests: number,
+  customerId: string
+}
